Fix thumbnail selector issue
the :host selector points to an Angular host element, however SASS required a slash to be added due to the line beginning with a colon; SCSS does not. This was causing an issue on production where thumbnails took up the full screen width.

Interestingly enough, the staging server does not show this behaviour.