Remove dynamic projects list from top page
The projects list updates far less frequently than the images, and judging by interest and word-of-mouth, visitors appear to be more interested in the artworks than the code projects.

This change removes the code module from the top page, and instead replaces it with a simpler blurb and link to the code projects list.