Implement recipe summary view using gallery styling
This change refactors the gallery thumbnail styles to a pair of shared files - one for the container and one for the thumbnails. These files are then used by the gallery and the recipe summary view to show a grid of images.

Additionally, this refactoring has included some code modernisation - flexbox is now used to lay out the image grid.